Split Chromium and Google Chrome
Split Chromium and Google Chrome because we may want to use them
indepedetly.
