+++ locate_chromium.go
@@ -10,7 +10,7 @@ import (
 var ChromiumExecutable = LocateChromium
 
 // LocateChromium returns a path to the Chromium binary, or an empty string if
-// Chromium installation is not found.
+// the Chromium installation is not found.
 func LocateChromium() string {
 
 	// If env variable "LORCACHROME" specified and it exists
@@ -24,27 +24,18 @@ func LocateChromium() string {
 	switch runtime.GOOS {
 	case "darwin":
 		paths = []string{
-			"/Applications/Google Chrome.app/Contents/MacOS/Google Chrome",
-			"/Applications/Google Chrome Canary.app/Contents/MacOS/Google Chrome Canary",
 			"/Applications/Chromium.app/Contents/MacOS/Chromium",
-			"/usr/bin/google-chrome-stable",
-			"/usr/bin/google-chrome",
 			"/usr/bin/chromium",
 			"/usr/bin/chromium-browser",
 		}
 	case "windows":
 		paths = []string{
-			os.Getenv("LocalAppData") + "/Google/Chrome/Application/chrome.exe",
-			os.Getenv("ProgramFiles") + "/Google/Chrome/Application/chrome.exe",
-			os.Getenv("ProgramFiles(x86)") + "/Google/Chrome/Application/chrome.exe",
 			os.Getenv("LocalAppData") + "/Chromium/Application/chrome.exe",
 			os.Getenv("ProgramFiles") + "/Chromium/Application/chrome.exe",
 			os.Getenv("ProgramFiles(x86)") + "/Chromium/Application/chrome.exe",
 		}
 	default:
 		paths = []string{
-			"/usr/bin/google-chrome-stable",
-			"/usr/bin/google-chrome",
 			"/usr/bin/chromium",
 			"/usr/bin/chromium-browser",
 			"/snap/bin/chromium",
